+++ R/unnest.R
@@ -14,16 +14,16 @@
 #'   b = 11:20,
 #'   c = c(rep("a", 6), rep("b", 4)),
 #'   d = c(rep("a", 4), rep("b", 6))) %>%
-#'   group_nest.(c, d)
+#'   nest_by.(c, d)
 #'
 #' nested_df %>%
-#'   unnest_legacy.(data)
-unnest_legacy. <- function(.data, col = NULL) {
-  UseMethod("unnest_legacy.")
+#'   unnest.(data)
+unnest. <- function(.data, col = NULL) {
+  UseMethod("unnest.")
 }
 
 #' @export
-unnest_legacy..data.frame <- function(.data, col = NULL) {
+unnest..data.frame <- function(.data, col = NULL) {
   if (!is_tidytable(.data)) .data <- as_tidytable(.data)
 
   col <- enexpr(col)
@@ -50,6 +50,6 @@ unnest_legacy..data.frame <- function(.data, col = NULL) {
 }
 
 #' @export
-#' @rdname unnest_legacy.
-dt_unnest_legacy <- unnest_legacy.
+#' @rdname unnest.
+dt_unnest_legacy <- unnest.